Update RA8875.cpp
Renamed variables to proper names, ie. x1,y1 & x2,y2. Added some out of bound checks. Also fixed point truncations. Now you can draw a partial rectangle that is out of bounds and in bounds, and it will only draw the in bounds portion. If the in bounds portion just happens to be 1 pixel, it will use the drawPixel function because _line_addressing does not like only 1 pixel.